+++ lib/I2C/I2C.h
@@ -3,4 +3,4 @@
 
 void I2Cinit(void);
 void I2CWriteReg(int8_t chipadr, int8_t regadr, int8_t data);
-void I2CReadReg(int8_t chipadr, int8_t regadr, byte *temp);
+void I2CReadReg(int8_t chipadr, int8_t regadr, int32_t *temp);
